GPACMuxMP4: avoid unsigned (and the corresponding cast)

+++ src/lib_media/mux/gpac_mux_mp4.cpp
@@ -545,7 +545,7 @@ void GPACMuxMP4::closeFragment() {
 			auto const absTimeInTs = convertToTimescale(firstDataAbsTimeInMs, 1000, mediaTs) + curFragmentStartInTs;
 			auto const deltaRealTimeInMs = 1000 * (double)(getUTC() - Fraction(absTimeInTs, mediaTs));
 
-			auto const isSuspicious = deltaRealTimeInMs < 0 || deltaRealTimeInMs > curFragmentStartInTs || (int64_t)curFragmentDurInTs != fractionToTimescale(segmentDuration, mediaTs);
+			auto const isSuspicious = deltaRealTimeInMs < 0 || deltaRealTimeInMs > curFragmentStartInTs || curFragmentDurInTs != fractionToTimescale(segmentDuration, mediaTs);
 			log(isSuspicious ? Warning : Debug,
 			    "Closing MSS fragment with absolute time %s %s UTC and duration %s (timescale %s, time=%s, deltaRT=%s)",
 			    getDay(), getTimeFromUTC(), curFragmentDurInTs, mediaTs, absTimeInTs, deltaRealTimeInMs);
